+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Deprecated risky `WP_REDIS_SERIALIZER` configuration constant, use `WP_REDIS_IGBINARY` instead
 - Support `WP_REDIS_USERNAME` when using Predis
 - Show cache hit ratio decimal points in Admin Bar node
+- Obscure secrets when displaying `WP_REDIS_SERVERS`
 
 ## 2.3.0